+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added implementation for `compas_gmsh.models.Model.from_brep`.
+
 ### Changed
 
 ### Removed